WIP - moved S4L models and AnnualCostsLimitedController models
