--- app/features/frs/forms/AnnualCostsLimitedFormFactory.scala
@@ -1,34 +1,0 @@
-/*
- * Copyright 2017 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package forms.frs
-
-import forms.FormValidation.textMappingWithMessageArgs
-import models.view.frs.AnnualCostsLimitedView
-import play.api.data.Form
-import play.api.data.Forms.mapping
-
-object AnnualCostsLimitedFormFactory {
-
-  val RADIO_COST_LIMITED: String = "annualCostsLimitedRadio"
-
-  def form(msgArgs: Seq[Any] = Seq()): Form[AnnualCostsLimitedView] = {
-    Form(mapping(
-      RADIO_COST_LIMITED -> textMappingWithMessageArgs()(msgArgs)("frs.costsLimited").verifying(AnnualCostsLimitedView.valid)
-    )(AnnualCostsLimitedView.apply)(AnnualCostsLimitedView.unapply))
-  }
-
-}
