+++ app/models/S4LKey.scala
@@ -39,6 +39,4 @@ object S4LKey {
   implicit val vatLodgingOfficer: S4LKey[S4LVatLodgingOfficer] = S4LKey("VatLodgingOfficer")
   implicit val vatFinancials: S4LKey[S4LVatFinancials] = S4LKey("VatFinancials")
   implicit val sicAndCompliance: S4LKey[S4LVatSicAndCompliance] = S4LKey("VatSicAndCompliance")
-  implicit val vatFlatRateScheme: S4LKey[S4LFlatRateScheme] = S4LKey("VatFlatRateScheme")
-
 }
